read config toml is shared
