+++ dbuzzy/src/db/mod.rs
@@ -1,6 +1,5 @@
-mod config;
-mod connection;
-mod types;
+pub mod connection;
+pub mod types;
 
 pub type DbResult<T> = anyhow::Result<T, Error>;
