+++ duzzy-lib/Cargo.toml
@@ -7,3 +7,6 @@ authors.workspace = true
 [dependencies]
 anyhow.workspace = true
 crossterm.workspace = true
+
+toml = "0.8.14"
+serde = { version = "1.0.203", features = ["derive"] }
