+++ duzzy-lib/src/lib.rs
@@ -1,1 +1,4 @@
 pub mod event;
+mod utils;
+
+pub use utils::config_toml;
